+++ src/pytest_lock/parser_file/builder.py
@@ -2,6 +2,7 @@ from typing import Dict
 
 from pytest_lock.parser_file.base import ParserFile
 from pytest_lock.parser_file.json import ParserFileJson
+from pytest_lock.parser_file.pickle import ParserFilePickle
 
 
 class ParserFileBuilder:
@@ -15,9 +16,7 @@ class ParserFileBuilder:
     mapping: Dict[str, ParserFile]
 
     def __init__(self):
-        self.mapping = {
-            ".json": ParserFileJson(),
-        }
+        self.mapping = {".json": ParserFileJson(), ".pickle": ParserFilePickle()}
 
     def build(self, extension: str = ".json") -> ParserFile:
         """
@@ -29,6 +28,8 @@ class ParserFileBuilder:
         parser_file = self.mapping.get(extension)
 
         if parser_file is None:
-            raise ValueError("No parser_file file found")
+            # Todo : Add a custom exception
+            # Todo : Add test for check if exception is raised with invalid extension
+            raise ValueError(f"ParserFile not found for extension : '{extension}'")
 
         return parser_file