+++ src/pytest_lock/plugin.py
@@ -7,28 +7,61 @@ Pytest will use the fixture 'lock' to give access to the lock feature.
 """
 
 from pathlib import Path
-from typing import Optional
+from typing import List, Optional
 
 import pytest
 from _pytest.config import Config
+from _pytest.python import Function
 from pytest import FixtureRequest
 
 from pytest_lock.cache import CacheLock
 from pytest_lock.config import ArgumentCLI, LockConfig
 from pytest_lock.fixture import FixtureLock
-from pytest_lock.parser_file.builder import ParserFileBuilder
+from pytest_lock.hooks.lock import skip_test_without_lock_fixture
+from pytest_lock.models.exceptions import LockCLIException
 
 CACHE_LOCK_PATH = ".pytest_lock"
 TESTS_PATH = "tests"
-EXTENSION = ".json"
+EXTENSION = ".pickle"
 
 
 def pytest_addoption(parser: Config):
     """Add new argument CLI to pytest."""
-    parser.addoption(ArgumentCLI.LOCK, action="store_true", help="Activate lock feature")  # type: ignore
+    parser.addoption(ArgumentCLI.LOCK, action="store_true", help="Activate lock feature, this argument make target only tests with 'lock' fixture.")  # type: ignore
     parser.addoption(ArgumentCLI.SIMULATE, action="store_true", help="Simulate lock feature")  # type: ignore
     parser.addoption(ArgumentCLI.LOCK_DATE, action="store", type=str, help="Activate lock date feature")  # type: ignore
     parser.addoption(ArgumentCLI.ONLY_SKIP, action="store_true", help="Lock only tests without lock")  # type: ignore
+    parser.addoption(ArgumentCLI.CLEAN, action="store_true", help="Clean lock file")  # type: ignore
+
+
+def check_cli_arguments(pytestconfig: Config) -> None:
+    """
+    Check if have error in cli arguments.
+
+    Args:
+        pytestconfig (Config): Pytest configuration.
+    """
+    is_lock = pytestconfig.getoption(ArgumentCLI.LOCK)
+    is_simulate = pytestconfig.getoption(ArgumentCLI.SIMULATE)
+    is_lock_date = pytestconfig.getoption(ArgumentCLI.LOCK_DATE)
+    ask_only_skip = pytestconfig.getoption(ArgumentCLI.ONLY_SKIP)
+    ask_clean = pytestconfig.getoption(ArgumentCLI.CLEAN)
+
+    if not is_lock:
+        if is_simulate:
+            raise LockCLIException(f"Can't activate '{ArgumentCLI.SIMULATE}' mode without '{ArgumentCLI.LOCK}'")
+        if is_lock_date:
+            raise LockCLIException(f"Can't activate '{ArgumentCLI.LOCK_DATE}', '{is_lock_date}' mode without '{ArgumentCLI.LOCK}'")
+        if ask_only_skip:
+            raise LockCLIException(f"Can't activate '{ArgumentCLI.ONLY_SKIP}' mode without '{ArgumentCLI.LOCK}'")
+        if ask_clean:
+            raise LockCLIException(f"Can't activate '{ArgumentCLI.CLEAN}' mode without '{ArgumentCLI.LOCK}'")
+
+    if ask_clean:
+        if is_lock_date:
+            raise LockCLIException(f"Can't activate '{ArgumentCLI.LOCK_DATE}', '{is_lock_date}' mode with '{ArgumentCLI.CLEAN}'")
+        if ask_only_skip:
+            raise LockCLIException(f"Can't activate '{ArgumentCLI.ONLY_SKIP}' mode with '{ArgumentCLI.CLEAN}'")
 
 
 @pytest.fixture(scope="function")
@@ -94,11 +127,14 @@ def _lock(
         FixtureLock: Lock fixture.
 
     """
-    root_path = pytestconfig.rootpath
+    # Check if have error in cli arguments
+    check_cli_arguments(pytestconfig)
+
     # Use default values if not given
+    root_path = pytestconfig.rootpath
+    extension = extension or EXTENSION
     tests_path = tests_path or root_path / TESTS_PATH
     cache_path = cache_path or root_path / CACHE_LOCK_PATH
-    extension = extension or EXTENSION
 
     # Create configuration for lock fixture
     config = LockConfig.from_pytest_run(
@@ -110,11 +146,55 @@ def _lock(
     )
 
     # Create cache system
-    parser_file_builder = ParserFileBuilder()
-    parser_file = parser_file_builder.build(EXTENSION)
-    cache_system = CacheLock(config, parser_file)
+    cache_system = CacheLock(config)
 
     # Create lock fixture with configuration and cache system
     lock_fixture = FixtureLock(config, cache_system)
 
     return lock_fixture
+
+
+def pytest_collection_modifyitems(config: Config, items: List[Function], name_fixture="lock") -> None:
+    """
+    Hook to modify tests received by pytest.
+
+    Notes:
+        This function allows you to act on tests retrieved by Pytest
+        before they start running. This allows you to ignore certain
+        tests, for example
+
+    Args:
+        config: Pytest configuration object.
+        items: List of tests received by pytest.
+        name_fixture: Name of the fixture to check.
+    """
+    conditions = (
+        config.getoption(ArgumentCLI.LOCK),  # --lock
+        config.getoption(ArgumentCLI.CLEAN),  # --lock
+    )
+
+    if any(conditions):
+        skip_test_without_lock_fixture(config, items, name_fixture=name_fixture)
+
+
+def pytest_configure(config: Config) -> None:
+    """
+    Hook to configure pytest for disable creation of cache file.
+
+    Notes:
+        When you do "pytest --lock --clean", the cache file is replace
+        by a new empty file. This hook allows you to disable the creation
+        of the cache file when the "--lock" argument is activated.
+
+    Args:
+        config: Pytest configuration object.
+    """
+    # Todo : Create a acceptation test for this hook
+    # Todo : Create a test for this hook and see if pytest-cache is not created
+    if config.getoption(ArgumentCLI.LOCK):
+        # Disable cache file
+        def fake_cache(*args, **kwargs):
+            pass
+
+        config.cache.set = fake_cache  # type: ignore
+        config.cache.get = fake_cache  # type: ignore