+++ tests/integrations/test_lock_change_result.py
@@ -1,13 +1,15 @@
 from _pytest.pytester import Pytester
 
+from pytest_lock.models.cli_argument import ArgumentCLI
+
 
 def test_lock_change_result(pytester: Pytester) -> None:
     """Test the creation of a cache."""
     pytester.copy_example("conftest.py")
     old_test_path = pytester.copy_example("scenarios/lock_change_result/test_step_1.py")
 
-    result = pytester.runpytest("--lock")
-    result.assert_outcomes(skipped=1)
+    result = pytester.runpytest(ArgumentCLI.LOCK)
+    result.assert_outcomes(passed=1)
 
     # Now the test is lock, change with test_step_2.py and run pytest again
     new_test_path = pytester.copy_example("scenarios/lock_change_result/test_step_2.py")
@@ -23,8 +25,8 @@ def test_lock_change_result_type(pytester: Pytester) -> None:
     pytester.copy_example("conftest.py")
     old_test_path = pytester.copy_example("scenarios/lock_change_result_type/test_step_1.py")
 
-    result = pytester.runpytest("--lock")
-    result.assert_outcomes(skipped=1)
+    result = pytester.runpytest(ArgumentCLI.LOCK)
+    result.assert_outcomes(passed=1)
 
     # Now the test is lock, change with test_step_2.py and run pytest again
     new_test_path = pytester.copy_example("scenarios/lock_change_result_type/test_step_2.py")